MC-1 Refactoring code, add improvements and java docs.
